+++ src/main/java/io/github/thebusybiscuit/slimefun4/api/events/GEOResourceGenerationEvent.java
@@ -129,12 +129,13 @@ public class GEOResourceGenerationEvent extends Event {
         return biome;
     }
 
-    public HandlerList getHandlers() {
+    public static HandlerList getHandlerList() {
         return handlers;
     }
 
-    public static HandlerList getHandlerList() {
-        return handlers;
+    @Override
+    public HandlerList getHandlers() {
+        return getHandlerList();
     }
 
 }