+++ src/main/java/io/github/thebusybiscuit/slimefun4/api/events/ReactorExplodeEvent.java
@@ -47,12 +47,13 @@ public class ReactorExplodeEvent extends Event {
         return reactor;
     }
 
-    public HandlerList getHandlers() {
+    public static HandlerList getHandlerList() {
         return handlers;
     }
 
-    public static HandlerList getHandlerList() {
-        return handlers;
+    @Override
+    public HandlerList getHandlers() {
+        return getHandlerList();
     }
 
 }