+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/SlimefunRegistry.java
@@ -13,7 +13,9 @@ import java.util.concurrent.ConcurrentHashMap;
 
 import org.bukkit.Location;
 import org.bukkit.Server;
+import org.bukkit.entity.Entity;
 import org.bukkit.entity.EntityType;
+import org.bukkit.entity.Piglin;
 import org.bukkit.inventory.ItemStack;
 
 import io.github.thebusybiscuit.cscorelib2.collections.KeyMap;
@@ -69,6 +71,7 @@ public class SlimefunRegistry {
     private final Set<String> tickers = new HashSet<>();
     private final Set<SlimefunItem> radioactive = new HashSet<>();
     private final Set<String> activeChunks = ConcurrentHashMap.newKeySet();
+    private final Set<ItemStack> barterDrops = new HashSet<>();
 
     private final KeyMap<GEOResource> geoResources = new KeyMap<>();
 
@@ -195,8 +198,24 @@ public class SlimefunRegistry {
         return layouts.get(layout);
     }
 
+    /**
+     * This returns a {@link Map} connecting the {@link EntityType} with a {@link Set}
+     * of {@link ItemStack ItemStacks} which would be dropped when an {@link Entity} of that type was killed.
+     * 
+     * @return The {@link Map} of custom mob drops
+     */
     public Map<EntityType, Set<ItemStack>> getMobDrops() {
         return mobDrops;
+    }
+
+    /**
+     * This returns a {@link Set} of {@link ItemStack ItemStacks} which can be obtained by bartering
+     * with {@link Piglin Piglins}.
+     * 
+     * @return A {@link Set} of bartering drops
+     */
+    public Set<ItemStack> getBarteringDrops() {
+        return barterDrops;
     }
 
     public Set<SlimefunItem> getRadioactiveItems() {
@@ -266,7 +285,7 @@ public class SlimefunRegistry {
     public Map<String, ItemStack> getAutomatedCraftingChamberRecipes() {
         return automatedCraftingChamberRecipes;
     }
-    
+
     public boolean logDuplicateBlockEntries() {
         return logDuplicateBlockEntries;
     }