+++ CHANGELOG.md
@@ -25,13 +25,31 @@
 
 #### Additions
 * Added an option for Industrial Miners to mine Ancient Debris
+* Added a new language: Korean
+* (API) Added support for adding custom Piglin Barter drops
+* (API) Added BlockPlacerPlaceEvent
+* (API) Added ToolUseHandler
 
 #### Changes
 * Performance improvement for Programmable Android rotations
+* Removed Gravel -> Flint recipe from the Grind stone
+* Performance improvements for miner talismans
+* Performance improvements for idling Enhanced Furnaces when using Paper
+* Performance improvements for Rainbow Blocks
 
 #### Fixes
 * Fixed Programmable Androids rotating in the wrong direction
 * Fixed #2176
+* Fixed #2164
+* Fixed #2147
+* Fixed #2179
+* Fixed Reinforced Spawners not working sometimes
+* Fixed Explosive Pickaxe not handling normal Shulker boxes correctly
+* Fixed #2103
+* Fixed #2184
+* Fixed #2183
+* Fixed #2181
+* Fixed #2180
 
 ## Release Candidate 15 (01 Aug 2020)